Allow displaying a single service, support service availability

+++ main.py
@@ -1,6 +1,7 @@
 from flask import Flask, session, render_template, request, jsonify, redirect
 from database import Database
 from image_server import ImageServer
+from utils import availability_to_int, availability_to_list
 
 app = Flask(__name__)
 app.config.from_object('config')
@@ -68,13 +69,22 @@ def api_service_list(category="all"):
     return jsonify(json_services)
 
 @app.route("/service/list/")
-def service_list():
-    services = db.get_all_services()
-    json_services = [dict(service) for service in services]
-    for json_service in json_services:
-        images = db.get_images_by_service_id(json_service['id'])
+@app.route("/service/list/<int:service_id>")
+def service_list(service_id=None):
+    if service_id:
+        service = db.get_service_by_id(service_id)
+        images = db.get_images_by_service_id(service_id)
+        json_service = dict(service)
+        json_service["available"] = availability_to_list(json_service["availability"])
         json_service['images'] = [{'filenameOnServer': image['filenameOnServer'], 'filename': image['filename']} for image in images]
-    return render_template("service/list.html", services=json_services)
+        return render_template("service/display.html", service=json_service)
+    else:
+        services = db.get_all_services()
+        json_services = [dict(service) for service in services]
+        for json_service in json_services:
+            images = db.get_images_by_service_id(json_service['id'])
+            json_service['images'] = [{'filenameOnServer': image['filenameOnServer'], 'filename': image['filename']} for image in images]
+        return render_template("service/list.html", services=json_services)
 
 @app.route("/api/user/service/list/<string:status>")
 def api_user_service_list(status="active"):
@@ -105,26 +115,34 @@ def user_service_create():
             failure_msg = "Please provide all the required fields"
             return ("user/service/create.html", failure_msg)
         
+        availability = availability_to_int(request.form.keys())
         images = request.files.getlist("images")
         files = image_server.store_images(images)
-        db.add_service(session["username"], title, description, category, files)
+        db.add_service(session["username"], title, description, category, availability, files)
         return redirect('/user/service/list/active')
 
 @app.route("/user/service/delete/<int:service_id>")
 def user_service_delete(service_id):
-    result = db.get_service_by_id(session["username"], service_id)
-    images = db.get_images_by_service_id(service_id)
-    image_server.remove_images(images)
-    db.remove_service(session["username"], service_id)
-    return redirect(f'/user/service/list/{result["status"]}')
+    result = db.get_service_by_id(service_id)
+    if result["username"] == session["username"]:
+        images = db.get_images_by_service_id(service_id)
+        image_server.remove_images(images)
+        db.remove_service(session["username"], service_id)
+        return redirect(f'/user/service/list/{result["status"]}')
 
 @app.route("/user/service/update/<int:service_id>", methods = ["GET", "POST"])
 def user_service_update(service_id):
     if "username" not in session:
         return render_template("user/account/login.html")
+    
+    service = db.get_service_by_id(service_id)
+    if service["username"] != session["username"]:
+        return render_template("home.html")
+    
     if request.method == "GET":
-        service = db.get_service_by_id(session["username"], service_id)
-        return render_template("user/service/create.html", service = dict(service))
+        service = dict(service)
+        service["available"] = availability_to_list(service["availability"])
+        return render_template("user/service/create.html", service = service)
     else:
         old_images = db.get_images_by_service_id(service_id)
         image_server.remove_images(old_images)
@@ -132,9 +150,10 @@ def user_service_update(service_id):
         title = request.form.get("title")
         description = request.form.get("description")
         category = request.form.get("category")
+        availability = availability_to_int(request.form.keys())
         images = request.files.getlist("images")
         files = image_server.store_images(images)
-        result = db.update_service(session["username"], service_id, title, description, category, files)
+        result = db.update_service(session["username"], service_id, title, description, category, availability, files)
         return redirect(f'/user/service/list/{result["status"]}')
 
 @app.route("/user/service/pause/<int:service_id>")
